exclude service providers from arch linting
